==> python-codes/06_pricing.py <==
# -*- coding: utf-8 -*-
"""
Created on Mon Jul 29 17:14:47 2019

@author: smedh
"""

import pandas as pd

df = pd.read_csv (r'C:\Users\smedh\Documents\PGPBABI\Capstone\ws_python\Datasets\03_kmode_result.csv')
